JSファイルの追加
- いいねに関する処理を`article.js`に移動し`application.js`でインポートするように修正。
